🔧 fix: 2-5 type 파일 분리
